feat: new modules 2d/mask, 3d/lights, core/{animation_tracks,component_scheduler,scene_globals} + docs
- src/2d/mask.rs: mask component with stencil/stage support
- src/3d/lights.rs: directional/spot/point lights with clamp helpers
- src/core/animation_tracks.rs: TrackInterpolation + Vector3/Real/Color/etc.
- src/core/component_scheduler.rs: per-thread start/update/late-update queue
- src/core/scene_globals.rs: ambient/fog/shadows/skybox/postprocess bundle
- docs/: completeness-check, interface-parity-audit, symbol maps,
  usage, perf comparison reports
- examples/perf_math.rs: math hot-path microbenchmark
- scripts/: interface-symbol-map generator + cocos4 vec3 test comparator
  + node math bench + test runner

Clippy clean, 1937 tests pass under cargo test -- --test-threads=4

+++ src/agi_minigame/ai_engine.rs
@@ -238,8 +238,8 @@ impl DimensionGenerator {
             let ti = template_indices.remove(idx);
             let template = &rule_templates[ti];
             let mut params = ValueMap::new();
-            params.insert("intensity".to_string(), Value::Float(difficulty as f32));
-            params.insert("duration".to_string(), Value::Float((10.0 + self.rng.gen::<f32>() * 30.0) as f32));
+            params.insert("intensity".to_string(), Value::Float(difficulty as f64));
+            params.insert("duration".to_string(), Value::Float((10.0 + self.rng.gen::<f32>() * 30.0) as f64));
 
             rules.push(GeneratedRule {
                 rule_id: template.0.to_string(),
@@ -418,7 +418,7 @@ impl RuleComposer {
 
             let mut params = ValueMap::new();
             let intensity = if has_synergy { difficulty * 1.5 } else { difficulty };
-            params.insert("intensity".to_string(), Value::Float(intensity as f32));
+            params.insert("intensity".to_string(), Value::Float(intensity as f64));
 
             rules.push(GeneratedRule {
                 rule_id: id.clone(),
